오류 수정 및 메인페이지 폼 추가
