+++ src/styles/components/auth/UserPropfileForm.scss
@@ -8,7 +8,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  min-height: 530px;
+  min-height: 700px;
   .profile {
     width: 360px;
     min-width: 360px;
